pycache

+++ disaster_relief/views/comments.py
@@ -34,6 +34,7 @@ def comments():
     if 'username' not in flask.session:
         return flask.redirect(flask.url_for('login'))
     logname = flask.session.get('username')
+    print("logname ", logname)
     operation = request.form.get('operation')
     postid = request.form.get('postid')
     commentid = request.form.get('commentid')
@@ -52,8 +53,11 @@ def comments():
             FROM comments
             WHERE commentid = ?
         ''', (commentid,)).fetchone()
+        if commentowner['owner'] is None:
+            abort(403)
         if commentowner['owner'] != logname:
             abort(403)
+
         connection.execute('''
             DELETE FROM comments
             WHERE commentid = ?
